Small update Added a square world coordinate system with start (0,0) at the center of a screen Started using google's code style guidelines And some refactoring was made again...

+++ game/src/window.cpp
@@ -2,10 +2,10 @@
 
 #include <GLFW/glfw3.h>
 
-std::pair<int, int> Window::getMousePosition() {
+std::pair<float, float> Window::getMousePosition() {
     double xpos = 0, ypos = 0;
     glfwGetCursorPos(m_window, &xpos, &ypos);
-    return {static_cast<int>(xpos), static_cast<int>(ypos)};
+    return {xpos, ypos};
 }
 
 GLFWwindow *Window::createWindow(std::string name, int width, int height) {
